(#85) Adicionado os scripts de modulos em pastas e separados, mais organizados
