+++ src/content.js
@@ -4,3 +4,5 @@ const script = document.createElement('script');
     script.setAttribute("src", chrome.runtime.getURL('index.js'));
     const head = document.head || document.getElementsByTagName("head")[0] || document.documentElement;
     head.insertBefore(script, head.lastChild);
+
+
