+++ src/manifest.json
@@ -11,7 +11,7 @@
   "web_accessible_resources": [{
     "resources": ["options.html",
                   "index.js",
-                  "services.js"],
+                  "/modules/*"],
     "matches": ["<all_urls>"],
     "extension_ids": []
    }],
